Release v1.0

+++ build/clear_up_containers.sh
@@ -1,11 +1,32 @@
 #!/usr/bin/env bash
 
+
+clear_cache="false"
+
+while getopts 'c' flag; do
+  case "${flag}" in
+    c) clear_cache="true" ;;
+  esac
+done
+
+if [[ -e ~/docker.log ]]
+then
 docker stop $(cat ~/docker.log)
 docker rm $(cat ~/docker.log)
-docker image prune -f
 rm ~/docker.log
+fi
+
+
+if [[ ${clear_cache} = "true" ]]
+ then
+ echo "REMOVING CACHE"
+docker image prune -f
+fi
+
+
 echo
 echo "Images left (just for checking whether no trash containers and images left)"
 docker images
 echo
 docker ps -a
+exit 0
